Updates how the settings rule validation works
